xray-agent/pyproject.toml
[tool.poetry]
name = "xray-agent"
version = "0.1.0"
description = "XRay Agent for VPN servers"
authors = ["HomeVPN Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
structlog = "^24.4.0"
httpx = "^0.27.2"
pyyaml = "^6.0.1"
cryptography = "^43.0.0"
grpcio = "^1.66.0"
grpcio-tools = "^1.66.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
black = "^24.8.0"
ruff = "^0.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"
